==> KML Challenge 2022F/3. catboost_9.0_820.66.ipynb <==
# Data Wrangling
import pandas as pd
from pandas import Series, DataFrame
import numpy as np

# Visualization
import matplotlib.pylab as plt
from matplotlib import font_manager, rc
import seaborn as sns
%matplotlib inline

# EDA
import klib

# Preprocessing & Feature Engineering
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import OrdinalEncoder
from sklearn.preprocessing import MinMaxScaler
from sklearn.impute import SimpleImputer
from sklearn.preprocessing import StandardScaler
from sklearn.preprocessing import MinMaxScaler
from sklearn.preprocessing import PowerTransformer
from sklearn.feature_selection import SelectPercentile
from sklearn.preprocessing import OrdinalEncoder, OneHotEncoder
from sklearn.pipeline import Pipeline
from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import FunctionTransformer
from sklearn.base import BaseEstimator, TransformerMixin
from sklearn.experimental import enable_iterative_imputer
from sklearn.impute import IterativeImputer

# Hyperparameter Optimization
from sklearn.model_selection import GridSearchCV
from sklearn.model_selection import RandomizedSearchCV

# Modeling
from sklearn import set_config
from catboost import CatBoostRegressor
from sklearn.linear_model import LinearRegression, Ridge, Lasso
import optuna
from optuna.distributions import CategoricalDistribution, IntDistribution, FloatDistribution
from optuna.integration import OptunaSearchCV, ShapleyImportanceEvaluator
from sklearn.model_selection import train_test_split, cross_val_score, cross_validate
from sklearn.model_selection import train_test_split, KFold
from sklearn.metrics import mean_squared_error


# Evaluation
from sklearn.model_selection import cross_val_score

# Utility
import os
import time
import random
import warnings; warnings.filterwarnings("ignore")
from IPython.display import Image
import pickle
from tqdm import tqdm
import platform
from itertools import combinations
from scipy.stats.mstats import gmean
from sklearn.feature_selection import RFE

import gc
import random
import re
from sklearn.metrics import *
from sklearn.model_selection import KFold
import warnings
warnings.filterwarnings(action='ignore')

tr0 = pd.read_csv('X_train.csv', encoding = 'cp949')
tr1 = pd.read_csv('gbm2_tr.csv', encoding = 'cp949')
tr2 = pd.read_csv('gbm3_tr.csv', encoding = 'cp949')
tr3 = pd.read_csv('gbm4_tr_fix.csv', encoding = 'cp949')
tr4 = pd.read_csv('gbm5_tr.csv', encoding = 'cp949')
tr5 = pd.read_csv('lm1_tr_fix.csv', encoding = 'cp949')
tr6 = pd.read_csv('lm2_tr.csv', encoding = 'cp949')
#tr6_2 = pd.read_csv('lm2_2_tr.csv', encoding = 'cp949')

te0 = pd.read_csv('X_test.csv', encoding = 'cp949')
te1 = pd.read_csv('gbm2_te.csv', encoding = 'cp949')
te2 = pd.read_csv('gbm3_te.csv', encoding = 'cp949')
te3 = pd.read_csv('gbm4_te_fix.csv', encoding = 'cp949')
te4 = pd.read_csv('gbm5_te.csv', encoding = 'cp949')
te5 = pd.read_csv('lm1_te_fix.csv', encoding = 'cp949')
te6 = pd.read_csv('lm2_te.csv', encoding = 'cp949')
#te6_2 = pd.read_csv('lm2_2_te.csv', encoding = 'cp949')

trj = pd.read_csv('은디디feature_selected_train.csv')
tej = pd.read_csv('은디디feature_selected_test.csv')

trc = pd.read_csv('채원만든피처_train.csv', encoding = 'cp949')
tec = pd.read_csv('채원만든피처_test.csv', encoding = 'cp949')

y_train = pd.read_csv('y_train.csv', encoding = 'cp949')

tr0.shape, tr1.shape, tr2.shape, tr3.shape, tr4.shape, tr5.shape, tr6.shape, trj.shape, trc.shape

te0.shape, te1.shape, te2.shape, te3.shape, te4.shape, te5.shape, te6.shape, tej.shape, tec.shape

### tr0/te0

tr0 = tr0.drop(['ID'],axis=1)
te0 = te0.drop(['ID'],axis=1)

numeric_features = ['대학성적']
categorical_features = ['직종','세부직종','직무태그','근무경력','근무형태','근무지역','출신대학','대학전공','어학시험','자격증']

tr0_num = tr0[numeric_features]
tr0_cat = tr0[categorical_features]

te0_num = te0[numeric_features]
te0_cat = te0[categorical_features]

print(len(tr0.columns))
print(len(te0.columns))
print(len(tr0_num.columns))
print(len(tr0_cat.columns))

### tr1/te1

#numeric_features = ['근무경력','근무경력_x_대학성적','대학_경력_standard deviation', '대학_경력_variance', '대학_경력_mean',
#                   '대학_경력_sum','대학성적']

#categorical_features = ['근무형태', '세부직종', '출신대학', '대학전공', '직무태그','직종', '인서울여부', '근무형태_계정(해)'
#                        , '근무지역' ,'근무지역_대분류', '어학시험', '근무지역_소분류']

tr1['자격증'] = tr1['자격증'].astype('object')
te1['자격증'] = te1['자격증'].astype('object')

tr1['인서울여부'] = tr1['인서울여부'].astype('object')
te1['인서울여부'] = te1['인서울여부'].astype('object')

tr1['큰손'] = tr1['큰손'].astype('object')
te1['큰손'] = te1['큰손'].astype('object')

tr1['사무직'] = tr1['사무직'].astype('object')
te1['사무직'] = te1['사무직'].astype('object')

tr1['직무수'] = tr1['직무수'].astype('object')
te1['직무수'] = te1['직무수'].astype('object')

numeric_features = tr1.select_dtypes('number').columns
categorical_features = tr1.select_dtypes('object').columns
#binary_features = ['자격증','인서울여부','큰손','사무직']

tr1_num = tr1[numeric_features]
tr1_cat = tr1[categorical_features]
#tr1_bin = tr1[binary_features]

te1_num = te1[numeric_features]
te1_cat = te1[categorical_features]
#te1_bin = te1[binary_features]

#tr1_cat = tr1_cat.drop(['자격증','인서울여부','큰손','사무직'], axis=1)
#te1_cat = te1_cat.drop(['자격증','인서울여부','큰손','사무직'], axis=1)

print(len(tr1.columns))
print(len(te1.columns))
print(len(tr1_num.columns))
print(len(tr1_cat.columns))
#print(len(tr1_bin.columns))

### tr2/te2

numeric_features = ['대학성적', '근무개월', '출신대학*대학성적']
categorical_features = ['직종', '세부직종', '직무태그', '근무형태','근무경력', '출신대학', '대학전공', '어학시험', '대학+전공', '직종+세부직종','근무형태+직종',
                        '자격증','충북', '전국', '전남', '경기', '기타', '일본', '울산', '강원', '충남',
       '미국', '광주', '경북', '전북', '해외', '경남', '중국', '인천', '제주', '캐나다', '서울', '부산',
       '대구', '대전', '해외기타', '신입경력']

tr2_num = tr2[numeric_features]
tr2_cat = tr2[categorical_features]
#tr2_bin = tr2[binary_features]

te2_num = te2[numeric_features]
te2_cat = te2[categorical_features]
#te2_bin = te2[binary_features]

print(len(tr2.columns))
print(len(te2.columns))
print(len(tr2_num.columns))
print(len(tr2_cat.columns))
#print(len(tr2_bin.columns))

### tr3/te3

#numeric_features = ['대학성적','근무경력_개월','근무경력_개월^2']
#categorical_features = ['직종_0','직종_1','직종_2','직종_3','세부직종','직무태그','근무경력','근무형태','출신대학','대학전공','어학시험','자격증','근무년수','장기근무','정규직_비정규직','여대','해외취업_유무','근무지역_0', '근무지역_1','근무지역_2','서울근무', '기타_모음','사무','it', '영업', '경영', '마케팅', '서비스', '디자인', '무역', '교육', '법', '기타']

del tr3['서울_교육']

del te3['과학']

#tr3 = tr3.replace('yes',1)
#tr3 = tr3.replace('no',0)

#tr3 = tr3.replace('O',1)
#tr3 = tr3.replace('X',0)

#te3 = te3.replace('yes',1)
#te3 = te3.replace('no',0)

#te3 = te3.replace('O',1)
#te3 = te3.replace('X',0)

#tr3 = tr3.replace({'장기근무':'장기근무 아님'},0)
#tr3 = tr3.replace({'장기근무':'장기근무'},1)

#te3 = te3.replace({'장기근무':'장기근무 아님'},0)
#te3 = te3.replace({'장기근무':'장기근무'},1)

#tr3 = tr3.replace({'정규직_비정규직':'비정규직'},0)
#tr3 = tr3.replace({'정규직_비정규직':'정규직'},1)

#te3 = te3.replace({'정규직_비정규직':'비정규직'},0)
#te3 = te3.replace({'정규직_비정규직':'정규직'},1)

#tr3 = tr3.replace('YES',1)
#tr3 = tr3.replace('NO',0)

#te3 = te3.replace('YES',1)
#te3 = te3.replace('NO',0)

numeric_features = tr3.select_dtypes('number').columns
categorical_features = tr3.select_dtypes('object').columns
#binary_features = ['장기근무', '정규직_비정규직',
#       '여대', '해외취업_유무', '서울근무', '기타_모음', '사무',
#       'it', '영업', '경영', '마케팅', '서비스', '디자인', '무역', '교육', '법', '기타',
#        '여대_마케팅', '서울_사무', '서울_it', '서울_영업', '서울_경영', '서울_마케팅',
#        '서울_서비스', '서울_디자인', '서울_무역', '서울_법']

tr3_num = tr3[numeric_features]
tr3_cat = tr3[categorical_features]
#tr3_bin = tr3[binary_features]

te3_num = te3[numeric_features]
te3_cat = te3[categorical_features]
#te3_bin = te3[binary_features]

#tr3_num = tr3_num.drop(['장기근무', '정규직_비정규직',
#       '여대', '해외취업_유무', '서울근무', '기타_모음', '사무',
#       'it', '영업', '경영', '마케팅', '서비스', '디자인', '무역', '교육', '법', '기타',
#        '여대_마케팅', '서울_사무', '서울_it', '서울_영업', '서울_경영', '서울_마케팅',
#        '서울_서비스', '서울_디자인', '서울_무역', '서울_법'], axis=1)
#te3_num = te3_num.drop(['장기근무', '정규직_비정규직',
#       '여대', '해외취업_유무', '서울근무', '기타_모음', '사무',
#       'it', '영업', '경영', '마케팅', '서비스', '디자인', '무역', '교육', '법', '기타',
#        '여대_마케팅', '서울_사무', '서울_it', '서울_영업', '서울_경영', '서울_마케팅',
#        '서울_서비스', '서울_디자인', '서울_무역', '서울_법'], axis=1)

print(len(tr3.columns))
print(len(te3.columns))
print(len(tr3_num.columns))
print(len(tr3_cat.columns))
#print(len(tr3_bin.columns))

### tr4/te4

del tr4['직종별태그여부']

del tr4['target']

numeric_features = tr4.select_dtypes('number').columns
categorical_features = tr4.select_dtypes('object').columns
#binary_features = ['경력유무', '근무형태_정규직', '근무형태_계약직',
#       '근무형태_해외취업', '근무형태_병역특례', '근무형태_인턴', '근무형태_파견직', '근무형태_unknown', '근무지역_수도권_서울', '근무지역_수도권_경기', '근무지역_수도권_인천',
#       '근무지역_강원', '근무지역_호남', '근무지역_호서', '근무지역_영남', '근무지역_제주', '근무지역_해외_유라시아',
#       '근무지역_해외_아메리카', '근무지역_해외_기타', '근무지역_기타', '어학시험여부', '경력여부']

tr4_num = tr4[numeric_features]
tr4_cat = tr4[categorical_features]
#tr4_bin = tr4[binary_features]

te4_num = te4[numeric_features]
te4_cat = te4[categorical_features]
#te4_bin = te4[binary_features]

#tr4_num = tr4_num.drop(['경력유무', '근무형태_정규직', '근무형태_계약직',
#       '근무형태_해외취업', '근무형태_병역특례', '근무형태_인턴', '근무형태_파견직', '근무형태_unknown', '근무지역_수도권_서울', '근무지역_수도권_경기', '근무지역_수도권_인천',
#       '근무지역_강원', '근무지역_호남', '근무지역_호서', '근무지역_영남', '근무지역_제주', '근무지역_해외_유라시아',
#       '근무지역_해외_아메리카', '근무지역_해외_기타', '근무지역_기타', '어학시험여부', '경력여부'], axis=1)
#te4_num = te4_num.drop(['경력유무', '근무형태_정규직', '근무형태_계약직',
#       '근무형태_해외취업', '근무형태_병역특례', '근무형태_인턴', '근무형태_파견직', '근무형태_unknown', '근무지역_수도권_서울', '근무지역_수도권_경기', '근무지역_수도권_인천',
#       '근무지역_강원', '근무지역_호남', '근무지역_호서', '근무지역_영남', '근무지역_제주', '근무지역_해외_유라시아',
#       '근무지역_해외_아메리카', '근무지역_해외_기타', '근무지역_기타', '어학시험여부', '경력여부'], axis=1)

print(len(tr4.columns))
print(len(te4.columns))
print(len(tr4_num.columns))
print(len(tr4_cat.columns))
#print(len(tr4_bin.columns))

### tr5/te5

numeric_features = ['대학성적','대학성적_nocl']
categorical_features = ['직종','세부직종','직무태그','근무경력','근무형태','출신대학','대학전공','어학시험','자격증','근무경력_y','근무경력_m','근무지역1','근무지역2','근무지역3','숙련도']

tr5_num = tr5[numeric_features]
tr5_cat = tr5[categorical_features]

te5_num = te5[numeric_features]
te5_cat = te5[categorical_features]

print(len(tr5.columns))
print(len(te5.columns))
print(len(tr5_num.columns))
print(len(tr5_cat.columns))

### tr6/te6

numeric_features = tr6.select_dtypes('number').columns
categorical_features = tr6.select_dtypes('object').columns

tr6_num = tr6[numeric_features]
tr6_cat = tr6[categorical_features]

te6_num = te6[numeric_features]
te6_cat = te6[categorical_features]

print(len(tr6.columns))
print(len(te6.columns))
print(len(tr6_num.columns))
print(len(tr6_cat.columns))

### trj/tej

numeric_features = trj.select_dtypes('number').columns
categorical_features = trj.select_dtypes('object').columns

trj_num = trj[numeric_features]
trj_cat = trj[categorical_features]

tej_num = tej[numeric_features]
tej_cat = tej[categorical_features]

print(len(trj.columns))
print(len(tej.columns))
print(len(trj_num.columns))
print(len(trj_cat.columns))

### trc/tec

numeric_features = trc.select_dtypes('number').columns
categorical_features = trc.select_dtypes('object').columns

trc_num = trc[numeric_features]
trc_cat = trc[categorical_features]

tec_num = tec[numeric_features]
tec_cat = tec[categorical_features]

print(len(trc.columns))
print(len(tec.columns))
print(len(trc_num.columns))
print(len(trc_cat.columns))

### concat

#tr_lst = [tr0,tr1,tr2,tr3,tr4,tr5,tr6]
#te_lst = [tr0,te1,te2,te3,te4,te5,te6]

#name = ['ORI', 'GBM2', 'GBM3', 'GBM4', 'GBM5', 'LM1', 'LM2']

#for df_lst in [tr_lst, te_lst]:
#  for i in range(len(df_lst)):
#    for j in range(len(df_lst[i].columns)):
#      df_lst[i].rename(columns = {df_lst[i].columns[j]:df_lst[i].columns[j]+'_'+name[i]}, inplace = True)

X_train_num = pd.concat([tr0_num,tr1_num,tr2_num,tr3_num,tr4_num,tr5_num,tr6_num,trj_num,trc_num], axis=1)
X_train_cat = pd.concat([tr0_cat,tr1_cat,tr2_cat,tr3_cat,tr4_cat,tr5_cat,tr6_cat,trj_cat,trc_cat], axis=1)
#X_train_bin = pd.concat([tr1_bin, tr2_bin, tr3_bin, tr4_bin], axis=1)

X_test_num = pd.concat([te0_num, te1_num,te2_num,te3_num,te4_num,te5_num,te6_num,tej_num,tec_num], axis=1)
X_test_cat = pd.concat([te0_cat, te1_cat,te2_cat,te3_cat,te4_cat,te5_cat,te6_cat,tej_cat,tec_cat], axis=1)
#X_test_bin = pd.concat([te1_bin, te2_bin, te3_bin, te4_bin], axis=1)

X_train_cat.shape, X_test_cat.shape, X_train_num.shape, X_test_num.shape

#X_train_bin.shape, X_test_bin.shape

#X_train_num.to_csv('X_train_num_notcleansing.csv', index = False, encoding = 'cp949')
#X_train_cat.to_csv('X_train_cat_notcleansing.csv', index = False, encoding = 'cp949')
#X_test_num.to_csv('X_test_num_notcleansing.csv', index = False, encoding = 'cp949')
#X_test_cat.to_csv('X_test_cat_notcleansing.csv', index = False, encoding = 'cp949')

### 중복 제거

df_num = pd.concat([X_train_num, X_test_num])
df_cat = pd.concat([X_train_cat, X_test_cat])
#df_bin = pd.concat([X_train_bin, X_test_bin])

df_num.shape, df_cat.shape#, df_bin.shape

#df_num.columns = list(pd.RangeIndex(len(df_num.columns)))
#df_cat.columns = list(pd.RangeIndex(len(df_cat.columns)))
#df_bin.columns = list(pd.RangeIndex(len(df_bin.columns)))

pd.set_option('display.max_seq_items', None)

df_num.columns

df_cat.columns

#del df_num['ID_LM1']

#del df_cat['ID_LM1']

df_num.columns = list(pd.RangeIndex(len(df_num.columns)))
df_cat.columns = list(pd.RangeIndex(len(df_cat.columns)))

df_num.columns

df_cat.columns

df_num = df_num.loc[:, ~df_num.T.duplicated()]
df_cat = df_cat.loc[:, ~df_cat.T.duplicated()]
#df_bin = df_bin.loc[:, ~df_bin.T.duplicated()]

df_num.columns

df_cat.columns

df_num.columns = list(pd.RangeIndex(len(df_num.columns)))
df_cat.columns = list(pd.RangeIndex(len(df_cat.columns)))
#df_bin.columns = list(pd.RangeIndex(len(df_bin.columns)))

df_num.columns

df_cat.columns

X_train_num = df_num.iloc[:16570,:]
X_test_num = df_num.iloc[16570:,:]
X_train_cat = df_cat.iloc[:16570,:]
X_test_cat = df_cat.iloc[16570:,:]
#X_train_bin = df_bin.iloc[:16570,:]
#X_test_bin = df_bin.iloc[16570:,:]

X_train_cat.shape, X_test_cat.shape, X_train_num.shape, X_test_num.shape#, X_train_bin.shape, X_test_bin.shape

X_train = pd.read_csv('X_train.csv', encoding='cp949').drop(columns='ID')
y_train = pd.read_csv('y_train.csv', encoding='cp949').Salary

X_test = pd.read_csv('X_test.csv', encoding='cp949')
test_id = X_test.ID
X_test = X_test.drop(columns='ID')

X_train_bin = X_train[['직종','세부직종','근무지역']]

X_test_bin = X_test[['직종','세부직종','근무지역']]

X_cat = pd.concat([X_train_cat, X_test_cat], axis=0)
X_num = pd.concat([X_train_num, X_test_num], axis=0)
X_bin = pd.concat([X_train_bin, X_test_bin], axis=0)

X_bin.columns

for col in X_cat.columns:
    X_cat.rename(columns = {col:str(col)+'cat'}, inplace = True)

for col in X_num.columns:
    X_num.rename(columns = {col:str(col)+'num'}, inplace = True)

for col in X_bin.columns:
    X_bin.rename(columns = {col:str(col)+'bin'}, inplace = True)

X_cat.columns

X_num.columns

X_bin.columns

X = pd.concat([X_cat, X_num, X_bin], axis = 1)

X.columns

X_train = X.iloc[:16570,:]
X_test = X.iloc[16570:,:]

numeric_features = X_num.columns.tolist() #X_num.columns
categorical_features = X_cat.columns.tolist() #X_cat.columns
binary_features = X_bin.columns.tolist() #X_bin.columns
#pca_features = X_pca.columns.tolist()

X_train = X_train[numeric_features+categorical_features+binary_features]  # 순서 주의!!!
X_test = X_test[numeric_features+categorical_features+binary_features]

cat_index = [list(X_train.columns).index(c) for c in categorical_features]

X_train.shape, X_test.shape

#pd.set_option('display.max_columns', None)

len(X_train.columns)

ori_xtrain = X_train.copy()

X_train.columns

X_train = X_train.drop([], axis=1)

X_test = X_test.drop([], axis=1)

from sklearn.pipeline import Pipeline, make_pipeline
from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import FunctionTransformer
from sklearn.base import BaseEstimator, TransformerMixin
from sklearn.impute import SimpleImputer
from sklearn.preprocessing import PowerTransformer
from sklearn.preprocessing import OrdinalEncoder, OneHotEncoder
from sklearn.feature_selection import SelectPercentile, SelectKBest
from sklearn.model_selection import train_test_split, KFold, cross_validate, ShuffleSplit
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.decomposition import PCA, TruncatedSVD
import statsmodels.api as sm

CATBOOST_VERSION = 9.0
NFOLDS = 14
SEED = 0
NCOMP = 31
P = 0.05

!pip install KMeans

from sklearn.cluster import KMeans

# 상하한값 제한을 통한 결측값 처리 함수: FunctionTransformer를 통해 호출
def remove_outlier(X, q=0.05):
    df = pd.DataFrame(X)
    return df.apply(lambda x: x.clip(x.quantile(q), x.quantile(1-q)), axis=0).values

# 회귀분석의 계수검정을 이용한 피처선택 전처리기 클래스
class MyFeatureSelector(TransformerMixin, BaseEstimator):
    # 전처리기 생성 즉, MyFeatureSelector() 호출시 실행
    def __init__(self, p=0.01):
        self.p = p

    # 전처리기의 fit() 호출시 실행
    def fit(self, X, y=None):
        X = sm.add_constant(X)
        results = sm.OLS(y, X).fit()
        self.cols = list(results.pvalues[1:] <= self.p)
        return self

    # 전처리기의 transform() 호출시 실행
    def transform(self, X):
        return X[:,self.cols].astype(np.int64)

numeric_transformer = Pipeline(
    steps=[
        ("imputer", SimpleImputer(strategy="mean")),
        ("outlier", FunctionTransformer(remove_outlier, kw_args={'q':0.05})),
    ]
)

categorical_transformer = Pipeline(
    steps=[
        ("imputer", SimpleImputer(strategy="most_frequent")),
        ("encoder", OrdinalEncoder(handle_unknown='use_encoded_value', unknown_value=99999, dtype=np.object)),
    ]
)

binary_transformer = Pipeline(
    steps=[
        ("impuer", FunctionTransformer(lambda x: x.fillna('없음'))),
        ("corpus", FunctionTransformer(lambda x: x.str.replace('·',',').str.split(',').str.join(" "))),
        ("BoW", CountVectorizer()),
        ("dense", FunctionTransformer(lambda x: x.toarray().astype(int), accept_sparse=True)),
    ]
)

column_transformer = ColumnTransformer(
    transformers=[
        ("num", numeric_transformer, numeric_features),
        ("cat", categorical_transformer, categorical_features),
        ("bin1", make_pipeline(binary_transformer, TruncatedSVD(n_components=NCOMP,random_state=SEED)), binary_features[0]),
        ("bin2", make_pipeline(binary_transformer, MyFeatureSelector(p=P)), binary_features[1]),
        ("bin3", make_pipeline(binary_transformer, MyFeatureSelector(p=P)), binary_features[2]),
    ]
)

preprocessor = Pipeline(
    steps=[
        ("column", column_transformer),
    ]
)

set_config(display="diagram")
preprocessor

X_train = preprocessor.fit_transform(X_train, y_train)
X_test = preprocessor.transform(X_test)

X = pd.DataFrame(X_train)
Y = pd.DataFrame(y_train)
Xt = pd.DataFrame(X_test)

kmeans = KMeans(n_clusters=5, init='k-means++')
kmeans.fit_transform(X,Y)
kmeans.transform(Xt)

pre_X5 = kmeans.predict(X)
k_means_5 = pd.DataFrame(pre_X5)
X['cluster5'] = k_means_5

X_train = np.array(X)
X_test = np.array(Xt)
y_train = np.array(Y)

%%time

# 최적화된 하이퍼파라미터로 OOF를 수행하여 최종 CatBoost 모형 생성:
# No tuning => tuning한 모델에 비해 성능이 떨어지지 않음

#sscv = ShuffleSplit(test_size=.3334, n_splits=5, random_state=0)
models = cross_validate(CatBoostRegressor(learning_rate = 0.05, cat_features=cat_index, verbose=True, random_state=SEED, n_estimators = 3000),
                        X_train, y_train,
                        cv=10,
                        scoring='neg_mean_squared_error',
                        return_estimator=True)
oof_pred = np.array([m.predict(X_test) for m in models['estimator']]).mean(axis=0)

scores = models['test_score']
print("\nCatBoost CV scores: ", np.sqrt(-1*scores))
print("CatBoost CV mean = %.2f" % np.sqrt(-1*scores.mean()), "with std = %.2f" % np.sqrt(scores.std()))

# submission 화일 생성
filename = f'catboost_{CATBOOST_VERSION}_{np.sqrt(-1*scores.mean()):.2f}.csv'
pd.DataFrame({'ID':test_id, 'Salary':oof_pred}).to_csv(filename, index=False)
