==> KML Challenge 2022F/4. weight_mean_ensemble.ipynb <==
## Creating mean ensembles from submission files

#### Weight

import pandas as pd
import numpy as np
import os
import warnings
warnings.filterwarnings("ignore")

folder = 'submissions'
nf = 0
for f in os.listdir(folder):
    ext = os.path.splitext(f)[-1]
    if ext == '.csv': 
        s = pd.read_csv(folder+"/"+f)
    else: 
        continue
    if nf == 0: 
        slist = s
    else: 
        slist = pd.merge(slist, s, on="ID")
    nf += 1

    
if nf >= 2:
    weight = [0.1, 0.1, 0.79, 0.01]
    model_num = [0, 1, 2, 3]

    pred_weight_results = [slist.iloc[:,j+1]*weight[j] for j in model_num]
    pred = (np.sum(np.array(pred_weight_results),axis = 0)) / (np.sum([weight[i] for i in model_num]))
    
    t = pd.Timestamp.now()
    fname = f"weight_mean_submission_{t.month:02}{t.day:02}_{t.hour:02}{t.minute:02}.csv"
    pd.DataFrame({'ID': slist.ID, 'salary': pred}).to_csv(fname, index=False)
    print(f"'{fname}' is ready to submit.")

## End
